+++ 2024-25-Islington-1stSemBlazorMAUI/Components/Pages/Index.razor
@@ -1,4 +1,2 @@
 ﻿@page "/"
 @using _2024_25_Islington_1stSemBlazorMAUI.Model
-@inject NavigationManager NavManager
-
