+++ 2024-25-Islington-1stSemBlazorMAUI/Components/_Imports.razor
@@ -8,3 +8,6 @@
 @using _2024_25_Islington_1stSemBlazorMAUI
 @using _2024_25_Islington_1stSemBlazorMAUI.Components
 @using _2024_25_Islington_1stSemBlazorMAUI.Services
+
+@*import NavigationManager*@
+@inject NavigationManager Nav
